split runtime-panel into components
