+++ ui/src/app/components/runtime-configuration-panel/disabled-panel.tsx
@@ -1,0 +1,28 @@
+import * as React from 'react';
+
+import { WarningMessage } from 'app/components/messages';
+import { TextColumn } from 'app/components/text-column';
+import { AoU } from 'app/components/text-wrappers';
+
+export const DisabledPanel = () => {
+  return (
+    <WarningMessage
+      data-test-id='runtime-disabled-panel'
+      iconSize={16}
+      iconPosition={'top'}
+    >
+      {
+        <TextColumn>
+          <div style={{ fontWeight: 600 }}>
+            Cloud services are disabled for this workspace.
+          </div>
+          <div style={{ marginTop: '0.5rem' }}>
+            You cannot run or edit notebooks in this workspace because billed
+            services are disabled for the workspace creator's <AoU /> Researcher
+            account.
+          </div>
+        </TextColumn>
+      }
+    </WarningMessage>
+  );
+};
