+++ ui/src/app/components/runtime-configuration-panel/gpu-config-selector.tsx
@@ -1,0 +1,105 @@
+import * as React from 'react';
+import { useEffect, useState } from 'react';
+import * as fp from 'lodash/fp';
+import { Dropdown } from 'primereact/dropdown';
+
+import { FlexColumn, FlexRow } from 'app/components/flex';
+import { CheckBox } from 'app/components/inputs';
+import { styles } from 'app/components/runtime-configuration-panel/styles';
+import { getValidGpuTypes, gpuTypeToDisplayName } from 'app/utils/machines';
+import { supportUrls } from 'app/utils/zendesk';
+
+export const GpuConfigSelector = ({
+  disabled,
+  onChange,
+  selectedMachine,
+  gpuConfig,
+}) => {
+  const { gpuType = 'nvidia-tesla-t4', numOfGpus = 1 } = gpuConfig || {};
+  const [selectedGpuType, setSelectedGpuType] = useState<string>(gpuType);
+  const [selectedNumOfGpus, setSelectedNumOfGpus] = useState<number>(numOfGpus);
+  const [hasGpu, setHasGpu] = useState<boolean>(!!gpuConfig);
+  const validGpuOptions = getValidGpuTypes(
+    selectedMachine.cpu,
+    selectedMachine.memory
+  );
+  const validGpuNames = fp.flow(
+    fp.map('name'),
+    fp.uniq,
+    fp.sortBy('price')
+  )(validGpuOptions);
+  const validNumGpusOptions = fp.flow(
+    fp.filter({ type: selectedGpuType }),
+    fp.map('numGpus')
+  )(validGpuOptions);
+
+  useEffect(() => {
+    onChange(
+      hasGpu && validGpuOptions.length > 0
+        ? {
+            gpuType: selectedGpuType,
+            numOfGpus: selectedNumOfGpus,
+          }
+        : null
+    );
+  }, [hasGpu, selectedGpuType, selectedNumOfGpus]);
+
+  return (
+    <FlexColumn style={styles.gpuSection}>
+      <FlexRow style={styles.gpuCheckBoxRow}>
+        <CheckBox
+          id={'enable-gpu'}
+          label='Enable GPUs'
+          checked={hasGpu}
+          style={styles.gpuCheckBox}
+          disabled={disabled}
+          onChange={() => {
+            setHasGpu(!hasGpu);
+          }}
+        />
+        <a target='_blank' href={supportUrls.gpus}>
+          Learn more about GPU cost and restrictions.
+        </a>
+      </FlexRow>
+      {hasGpu && (
+        <FlexRow style={styles.formGrid2}>
+          <FlexRow style={styles.labelAndInput}>
+            <label
+              style={{ ...styles.label, minWidth: '3.0rem' }}
+              htmlFor='gpu-type'
+            >
+              Gpu Type
+            </label>
+            <Dropdown
+              id={'gpu-type'}
+              style={{ width: '7rem' }}
+              options={validGpuNames}
+              onChange={({ value }) => {
+                setSelectedGpuType(
+                  fp.find({ name: value }, validGpuOptions).type
+                );
+              }}
+              disabled={disabled}
+              value={gpuTypeToDisplayName(selectedGpuType)}
+            />
+          </FlexRow>
+          <FlexRow style={styles.labelAndInput}>
+            <label
+              style={{ ...styles.label, minWidth: '2.0rem' }}
+              htmlFor='gpu-num'
+            >
+              GPUs
+            </label>
+            <Dropdown
+              id={'gpu-num'}
+              options={validNumGpusOptions}
+              onChange={({ value }) => setSelectedNumOfGpus(value)}
+              disabled={disabled}
+              value={selectedNumOfGpus}
+            />
+          </FlexRow>
+        </FlexRow>
+      )}
+    </FlexColumn>
+  );
+};
